+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Make TheAlgorithm.weightedScorers private
 * Make `isFollowed` and `isFollower` actually optional on `Account`
 
 ### v0.69.17